feat(nexus_utils): add Nexus utility functions
Added new utility functions for handling Nexus repository operations, including URL splitting, authentication header creation, and component management. Introduced custom exceptions for better error handling. Refactored existing code to utilize these utilities, improving modularity and reusability.

- Added `split_repository_url` to parse repository URLs.
- Implemented `create_auth_headers` for generating request headers.
- Developed `get_repository_details` to fetch repo format/type.
- Created `check_component_exists` to verify component presence.
- Added `delete_component_by_id` for component deletion by ID.
- Moved common logic from modules to shared utils.

BREAKING CHANGE: Refactored module imports; update import paths in dependent modules.
